Add files via upload
This is the Jupyter notebook file in which all the manipulation work with help of pandas which is a library of the Python is done. Here various column where created and added like moving average from 50 and 200 and many more columns like previous day volumes , closing day price and etc.